+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Improvements
 
 - Added `GicDistributor::pidr2()` and `GicRedistributor::pidr2()` methods.
+- Added example use of `GicV3` with QEMU's `virt` machine under `examples/`.
 
 ## 0.8.1